pyproject.toml
[tool.poetry]
name = "solutions"
version = "0.1.0"
description = ""
authors = ["geniungamartya <geniungamartya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.17"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.19"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
